refactor: report retrieve 관심사 분리
