+++ src/test/java/bsise/server/report/retrieve/service/ReportStatusRetrieveServiceTest.java
@@ -6,13 +6,13 @@ import static org.junit.jupiter.api.Assertions.*;
 import bsise.server.auth.OAuth2Provider;
 import bsise.server.letter.Letter;
 import bsise.server.letter.LetterRepository;
-import bsise.server.report.CoreEmotion;
-import bsise.server.report.DailyReport;
-import bsise.server.report.WeeklyReport;
-import bsise.server.report.daily.DailyReportRepository;
+import bsise.server.report.daily.domain.CoreEmotion;
+import bsise.server.report.daily.domain.DailyReport;
+import bsise.server.report.weekly.domain.WeeklyReport;
+import bsise.server.report.daily.repository.DailyReportRepository;
 import bsise.server.report.retrieve.dto.DailyReportStatusResponseDto;
 import bsise.server.report.retrieve.dto.WeeklyReportStatusResponseDto;
-import bsise.server.report.weekly.WeeklyReportRepository;
+import bsise.server.report.weekly.repository.WeeklyReportRepository;
 import bsise.server.user.domain.Preference;
 import bsise.server.user.domain.Role;
 import bsise.server.user.domain.User;
